bq-pagination/requirements.txt
google-cloud-bigquery>=3.25.0
fastapi[standard]>=0.115.5